+++ app/femr/ui/models/admin/inventory/InventoryViewModelPost.java
@@ -29,7 +29,7 @@ public class InventoryViewModelPost {
     private Integer medicationQuantity;
     private String medicationForm;
     private String medicationName;
-    private List<Integer> medicationStrength;
+    private List<Double> medicationStrength;
     private List<String> medicationUnit;
     private List<String> medicationIngredient;
 
@@ -89,11 +89,11 @@ public class InventoryViewModelPost {
     }
 
 
-    public List<Integer> getMedicationStrength() {
+    public List<Double> getMedicationStrength() {
         return medicationStrength;
     }
 
-    public void setMedicationStrength(List<Integer> medicationStrength) {
+    public void setMedicationStrength(List<Double> medicationStrength) {
         this.medicationStrength = medicationStrength;
     }
